fix
(cherry picked from commit 9449d4649ab06437ec48f9ba20aac478c685e44a)

+++ calvalus-processing/src/main/java/com/bc/calvalus/processing/fire/format/grid/AbstractGridReducer.java
@@ -238,7 +238,7 @@ public abstract class AbstractGridReducer extends Reducer<Text, GridCells, NullW
         for (int x = 1; x < numRowsGlobal; x++) {
             array[x] = array[x - 1] - 180.0f / numRowsGlobal;
         }
-        Array values = Array.factory(DataType.FLOAT, new int[]{numRowsGlobal}, array);
+        Array values = Array.factory(DataType.DOUBLE, new int[]{numRowsGlobal}, array);
         ncFile.write(lat, values);
     }
